Migrated signin to JS, fixed bugs
