+++ frontend/src/pages/App.jsx
@@ -1,7 +1,4 @@
-import React, { useState } from 'react';
-import SignIn from './components/SignIn';
-import SignUp from './components/SignUp';
-import './styles/App.css';
+import React from 'react';
 
 function App() {
