+++ frontend/src/styles/About.css
